+++ src/lib/app-fns/tx/common/catchTxError.tsx
@@ -36,7 +36,6 @@ export const catchTxError = (
 ): OperatorFunction<TxResultRendering, TxResultRendering> => {
   return catchError((error: Error) => {
     const txHash = error.message.match("(?:tx )(.*?)(?= at)")?.at(1);
-
     AmpTrack(
       error.message === "Request rejected"
         ? AmpEvent.TX_REJECTED